DBChecker: reconcile the election timeouts when specified
We may want to update the ovn election timers, and in order to do that
we need to act asynchronously because the election must be completed and
the change must be performed on the leader node. So, we add two optional
parameters for the desired election timer and we change that as part of
the db checker timer.

Signed-off-by: Federico Paolinelli <fpaoline@redhat.com>